+++ libc/string/strlen.c
@@ -1,7 +1,7 @@
 #include "../include/string.h"
 #include <stddef.h>
 
-int strlen(const char* str) {
+size_t strlen(const char* str) {
     size_t len = 0;
 
     while(str[len])
